+++ packages/kad-dht/test/rpc/index.node.ts
@@ -1,6 +1,6 @@
 /* eslint-env mocha */
 
-import { EventEmitter } from '@libp2p/interface/events'
+import { TypedEventEmitter } from '@libp2p/interface/events'
 import { start } from '@libp2p/interface/startable'
 import { mockStream } from '@libp2p/interface-compliance-tests/mocks'
 import { PersistentPeerStore } from '@libp2p/peer-store'
@@ -51,7 +51,7 @@ describe('rpc', () => {
     }
     components.peerStore = new PersistentPeerStore({
       ...components,
-      events: new EventEmitter<Libp2pEvents>()
+      events: new TypedEventEmitter<Libp2pEvents>()
     })
 
     await start(...Object.values(components))